Upload the working file under src
