+++ Project3/src/CertificateStore.java
@@ -1,15 +1,20 @@
+package working;
 
 
 /*
 EMU COSC 480/592
 PKI Project
 This class holds an arraylist of all certificates.
-Initially, we will hard code certificates for the group members. Once the program works with these, 
-we will add the ability for users to request/create certificates and other functionality.
-*/
+Initially, we will hard code certificates for the group members using the constructor.
+Later, we can add file IO capabilities to store & retrieve certificates.
+ */
 import java.util.ArrayList;
 
+
 public class CertificateStore {
+	
+	
+	
 
 //	private ArrayList<Certificate> certs = new ArrayList<Certificate>();
 	private ArrayList<Certificate> certs;
@@ -20,12 +25,12 @@ public class CertificateStore {
 	
 	public void createStore() {
 //		All certificates but Brain are valid with an ultra super top secret private key: "privateKey"
-		this.certs.add(new Certificate("2021-12-31", "Brian", "DH", "These Params", "privateKey"));
-		this.certs.add(new Certificate("2021-01-01", "Brain", "DH", "These Params", "privateKey"));
-		this.certs.add(new Certificate("2021-12-31", "Sarah", "DH", "These Params", "privateKey"));
-		this.certs.add(new Certificate("2021-12-31", "Payton", "DH", "These Params", "privateKey"));
-		this.certs.add(new Certificate("2021-12-31", "Fatema", "DH", "These Params", "privateKey"));
-		this.certs.add(new Certificate("2021-12-31", "Yu", "DH", "These Params", "privateKey"));
+		this.certs.add(new Certificate("2021-12-31", "Brian", "DH", "These Params"));
+		this.certs.add(new Certificate("2021-01-01", "Brain", "DH", "These Params"));
+		this.certs.add(new Certificate("2021-12-31", "Sarah", "DH", "These Params"));
+		this.certs.add(new Certificate("2021-12-31", "Payton", "DH", "These Params"));
+		this.certs.add(new Certificate("2021-12-31", "Fatema", "DH", "These Params"));
+		this.certs.add(new Certificate("2021-12-31", "Yu", "DH", "These Params"));
 		
 	}
 	
@@ -34,6 +39,7 @@ public class CertificateStore {
 	public ArrayList<Certificate> getCerts() {
 		return certs;
 	}
+	
 
 	public void setCerts(ArrayList<Certificate> certs) {
 		this.certs = certs;
@@ -64,7 +70,8 @@ public class CertificateStore {
 		
 		for (Certificate c : this.certs) {
 			if (c.userAndKeyMatch(user, userKey)) {
-				return(c.dateIsValid());
+				
+				return(c.dateIsValid() && c.validatePublicKey(c.getUserPublicKey()));
 			}		
 		}
 		System.out.println("Error: No match for user and key found.");
