+++ Project3/src/sy_chat.txt
@@ -1,3 +1,3 @@
-Brain has joined the chat!
-Brain has left the chat.
-Brain has left the chat.
+brian has joined the chat!
+brian: Hey
+brian has left the chat.
